+++ NEWS.md
@@ -1,8 +1,8 @@
 # ODRF 0.0.3
 
 * The function predicate.ODT() runs error when ODT is not split (depth=1), and we have fixed this bug.
-* We have fixed the functions VarImp() and plot.VarImp().
 * We have fixed the function predict.ODRF with arguments numOOB and weight.tree related issues.
+* We have fixed the functions VarImp() and plot.VarImp().
 
 # ODRF 0.0.2